requirements.txt
docopt==0.6.2
numpy==1.11.0
Pillow==3.2.0
